requirements.txt
# Core data processing
polars>=0.20.0
duckdb>=0.9.0
numpy>=1.26.0
scipy>=1.11.0

# ML / Statistics
xgboost>=2.0.0
lightgbm>=4.1.0
scikit-learn>=1.3.0
arch>=6.2.0
statsmodels>=0.14.0

# Technical analysis
ta>=0.11.0

# Configuration
pyyaml>=6.0

# Data fetching (placeholder - will use ccxt later)
yfinance>=0.2.30

# Utilities
click>=8.1.0
rich>=13.0.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
